+++ CHANGELOG.md
@@ -7,6 +7,7 @@ SMARTGEAR Project for Contao Open Source CMS
 - Smartgear : Dashboard now show informations about all root pages domains instead of the one filled in Smartgear's configuration
 - Smartgear : Users have to select an image size for contents, unless they are granted the expert role
 - Smartgear : FileUsage bundle configuration updated
+- Smartgear : Adding `##useful_data##` & `##useful_data_filled##` tokens for forms' notifications
 - FormContact : Sender's address is now `##admin_email##` token instead of Smartgear's Owner' email address
 
 1.0.20 - 2023-08-04