+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 SMARTGEAR Project for Contao Open Source CMS
 ========
 
-1.0.21 - 2023-08-08
+1.0.21 - 2023-08-09
 - Smartgear : Smartgear's API `version` endpoint now returns PHP, Contao `core-bundle` & Framway's version
 - Smartgear : Do not register visits if not pointing on a page
 - Smartgear : Dashboard now show informations about all root pages domains instead of the one filled in Smartgear's configuration